SPT-1514: Split handlers directory into domain specific directories
The `types` directory has a big mish-mash of things in it - some domain
logic and some more common domain-independent logic. This commit aims
to tidy up our types by putting them closer to the code that uses them.

The `handlers` directory currently contains code for two different bits
of functionality - handling AIS messages and responding to the
/user-identity endpoint on the API. This commit splits that handler
code into two sub-directories, one for ais-message-processor and one
for user-identity. The first has been renamed from
`txma-message-processor` because the word `txma` can be confused with
producing (like audit) and consuming messages. It felt clearer to add
"ais" instead of txma, as code here is responsible for responding to
AIS messages.

Types from `interfaces.ts` that are related to the input/output data
transfer for those handlers has also been split out and co-located with
the relevant handlers.
