+++ src/handlers/ais-message-processor/ais-message.ts
@@ -1,0 +1,10 @@
+import { isStringWithLength } from "../../types/string-utils";
+
+export type AisMessage = {
+  user_id: string;
+  timestamp: number;
+  intervention_code?: string;
+};
+
+export const isAisMessage = (request: Record<string, string | number>): request is AisMessage =>
+  request && isStringWithLength(request.user_id) && typeof request.timestamp === "number";
