+++ src/handlers/user-identity/tests/user-identity-handler.test.ts
@@ -1,10 +1,10 @@
 import { APIGatewayProxyEvent, Context } from "aws-lambda";
 import { handler } from "../user-identity-handler";
-import { StoredIdentityResponse } from "../../types/interfaces";
-import { HttpCodesEnum } from "../../types/constants";
-import { Configuration } from "../../types/configuration";
-import * as configuration from "../../types/configuration";
-import { CredentialStoreIdentityResponse } from "../../credential-store/credential-store-identity-response";
+import { HttpCodesEnum } from "../../../types/constants";
+import { Configuration } from "../../../types/configuration";
+import * as configuration from "../../../types/configuration";
+import { CredentialStoreIdentityResponse } from "../../../credential-store/credential-store-identity-response";
+import { UserIdentityResponseMetadata } from "../user-identity-response-metadata";
 
 describe("user-identity-handler tests", () => {
   const event = () => {
@@ -53,7 +53,7 @@ describe("user-identity-handler tests", () => {
     const result = await handler(newEvent, {} as Context);
 
     expect(result.statusCode).toBe(HttpCodesEnum.OK);
-    const body = JSON.parse(result.body) as StoredIdentityResponse;
+    const body = JSON.parse(result.body) as UserIdentityResponseMetadata;
     expect(body.vot).toBe("P2");
     expect(body.content).toEqual({
       sub: "user-sub",
