+++ src/handlers/user-identity/user-identity-response-metadata.ts
@@ -1,0 +1,10 @@
+import { UserIdentityResponse } from "./user-identity-response";
+
+export interface UserIdentityResponseMetadata {
+  content: UserIdentityResponse;
+  isValid: boolean;
+  expired: boolean;
+  vot: string;
+  kidValid: boolean;
+  signatureValid: boolean;
+}
