+++ src/handlers/user-identity/user-identity-response.ts
@@ -1,0 +1,5 @@
+export interface UserIdentityResponse {
+  sub: string;
+  vot: string;
+  vtm: string[];
+}
